챌린지 비용 결제 기능

+++ src/main/java/com/growable/starting/dto/LectureDto.java
@@ -21,7 +21,7 @@ public class LectureDto {
     private LocalDate recruitmentStartDate;
     private LocalDate recruitmentEndDate;
     private int capacity;
-    private double fee;
+    private int fee;
     private LocalDate lectureStartDate;
     private LocalDate lectureEndDate;
     private String mentorName;
